Implement authorization and route protection

+++ src/main/java/com/ivanzkyanto/senopi/service/impl/NoteServiceImpl.java
@@ -2,6 +2,7 @@ package com.ivanzkyanto.senopi.service.impl;
 
 import com.ivanzkyanto.senopi.entity.Note;
 import com.ivanzkyanto.senopi.entity.Tag;
+import com.ivanzkyanto.senopi.entity.User;
 import com.ivanzkyanto.senopi.model.request.CreateNoteRequest;
 import com.ivanzkyanto.senopi.model.request.UpdateNoteRequest;
 import com.ivanzkyanto.senopi.model.response.NoteResponse;
@@ -33,15 +34,17 @@ public class NoteServiceImpl implements NoteService {
     public TagRepository tagRepository;
 
     private final ResponseStatusException NOTE_NOT_FOUND_EXCEPTION = new ResponseStatusException(HttpStatus.NOT_FOUND, "Catatan tidak ditemukan");
+    private final ResponseStatusException NOTE_FORBIDDEN_EXCEPTION = new ResponseStatusException(HttpStatus.FORBIDDEN, "Anda tidak berhak mengakses resource ini");
 
     @Override
     @Transactional
-    public String create(CreateNoteRequest request) {
+    public String create(User user, CreateNoteRequest request) {
         validationService.validatePayload(request);
 
         Note note = Note.builder()
                 .title(request.title())
                 .body(request.body())
+                .owner(user)
                 .build();
 
         noteRepository.save(note);
@@ -59,27 +62,29 @@ public class NoteServiceImpl implements NoteService {
     }
 
     @Override
-    public List<NoteResponse> getAll() {
-        List<Note> notes = noteRepository.findAll();
+    public List<NoteResponse> getAll(User user) {
+        List<Note> notes = noteRepository.findAllByOwner(user);
 
         return notes.stream().map(NoteResponse::buildFrom).toList();
     }
 
     @Override
-    public NoteResponse get(String noteId) {
+    public NoteResponse get(User user, String noteId) {
         UUID uuid = validationService.validateUuid(noteId, () -> NOTE_NOT_FOUND_EXCEPTION);
         Note note = noteRepository.findById(uuid).orElseThrow(() -> NOTE_NOT_FOUND_EXCEPTION);
+        if (!note.getOwner().getId().equals(user.getId())) throw NOTE_FORBIDDEN_EXCEPTION;
 
         return NoteResponse.buildFrom(note);
     }
 
     @Override
     @Transactional
-    public void update(String noteId, UpdateNoteRequest request) {
+    public void update(User user, String noteId, UpdateNoteRequest request) {
         UUID uuid = validationService.validateUuid(noteId, () -> NOTE_NOT_FOUND_EXCEPTION);
         validationService.validatePayload(request);
 
         Note note = noteRepository.findById(uuid).orElseThrow(() -> NOTE_NOT_FOUND_EXCEPTION);
+        if (!note.getOwner().getId().equals(user.getId())) throw NOTE_FORBIDDEN_EXCEPTION;
 
         note.setTitle(request.title());
         note.setBody(request.body());
@@ -94,9 +99,10 @@ public class NoteServiceImpl implements NoteService {
 
     @Override
     @Transactional
-    public void delete(String noteId) {
+    public void delete(User user, String noteId) {
         UUID uuid = validationService.validateUuid(noteId, () -> NOTE_NOT_FOUND_EXCEPTION);
         Note note = noteRepository.findById(uuid).orElseThrow(() -> NOTE_NOT_FOUND_EXCEPTION);
+        if (!note.getOwner().getId().equals(user.getId())) throw NOTE_FORBIDDEN_EXCEPTION;
 
         noteRepository.delete(note);
     }
